requirements.txt
black==22.6.0
click==8.1.3
flake8==4.0.1
llvmlite==0.38.1
mccabe==0.6.1
mypy-extensions==0.4.3
numba==0.55.2
numpy==1.22.4
opencv-python==4.6.0.66
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
tomli==2.0.1
